Placeholder refactoring to prepare for the use of dynamic customization by the users, such as customizing each and every bar graph, or number formatting, where each use of the same placeholder could have different settings.
